barista-agent-system/pyproject.toml
[tool.poetry]
name = "barista-agent-system"
version = "0.1.0"
description = ""
authors = ["jggomez <juan.gomez01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-adk = "^1.17.0"
google-cloud-firestore = "^2.21.0"
pillow = "^12.0.0"
python-dotenv = "^1.2.1"
pyyaml = "^6.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
